log-generator/requirements.txt
flask==3.1.0
flask-limiter==3.8.0
pandas==2.2.3
requests==2.32.3
python-dotenv==1.0.1
gunicorn==21.2.0
